Sema: Fix crash with stored property that has on-demand 'modify' accessor
If a stored property has observers, the 'modify' coroutine body takes a
special form which directly yields the underlying storage and calls the
'didSet' and/or 'willSet' accessors.

However, if a property is 'dynamic', the 'modify' coroutine is synthesized
on-demand and it's SIL is serialized. It cannot reference the 'didSet'
or 'willSet' accessors, since they are private. Furthermore, an on-demand
accessor must also work if the property is overridden in a subclass.

So in this case, fall back to the more general 'modify' synthesis code,
which just calls the getter followed by the setter.

Fixes <rdar://problem/62339808>.