[Diags] Don't error twice for incomplete computed properties in extensions
If you have an empty computed property in an extension:

```swift
struct S {}

extension S {
    var foo: Int {
    }
}
```

Then Swift will give 2 conflicting errors: one that calls it a computed
property and another that calls it a stored property.

```
5:5: error: computed property must have accessors specified
    }
    ^
4:9: error: extensions must not contain stored properties
    var foo: Int {
        ^
```

This removes the unhelpful second diagnostic by adding a getter to the
property when the first diagnostic is in an extension.

This happens to me frequently. I'll decide to add a computed property,
write its declaration, then pause to decide how to implement it. The
presence of a 2nd error is both distracting and confusing.