runtime, build: homogenise the assertions toggle
Sink a couple of build configuration toggles related to the runtime into
the standard library build.

Unify the `SWIFT_STDLIB_ASSERTIONS` and `LLVM_ENABLE_ASSERTIONS` into a
new `SWIFT_RUNTIME_ENABLE_ASSERTIONS` which matches the style of the
other Swift runtime options.

Delay the configuration rendering until the end of the configuration to
allow the options to propagate back up.  This needs to be done until the
runtime configuration is fully sunk into the runtime build.

Remove the use of `LLVM_ENABLE_ASSERTIONS` in the runtime in preparation
to sever the LLVM dependency for the runtime, ingesting the LLVM Support
ADT definitions.  That will allow the ODR violations to be eradicated by
means of an inline namespace.