runtime: remove dependency on `LLVM_ENABLE_THREADS`
This removes the LLVMSupport usage for thread local.  This effectively
is a no-op as the runtime requires that it is built with clang, which
provides the C++11 `thread_local` support on all the targets.  In the
case that the target does not support `thread_local`, clang would
fallback to `__declspec(thread)` on MSVC and `__thread` on other modes.
However, because the runtime is always built in C++14 mode, the
`thread_local` support should be assumed to be present.  Remove the
unnecessary condition and inline the single use of the macro for
`thread_local`.