[5.3] IRGen: Only replace opaque result types of function result types when performing a dynamic replacement
Explanation: When a opaque result type (``some P``) appeared as a parameter of a
dynamic replacement function we used to try to emit this opaque result
type as a replacement. This made no sense since you can only replace
result types that are opaque result types not parameters that happen to
have the type of an opaque result type. And doing so would lead to a
compiler crash.

Scope: Any use of a dynamic replacement with an opaque result type
parameter would crash the compiler.

Origination: This bug was introduced together with the dynamic
replacement feature.

Risk: Medium. Affects only dynamic replacements with opaque result
types.

Reviewed by: Joe Groff

Testing: Unit test added.

rdar://61936622