Propagate `-vfsoverlay` from driver to frontend.
Today, the driver does not propagate the flag to the frontend
but also does not emit an error (silently consuming the flag
instead).

I'll open a similar PR for apple/swift-driver once this one is
merged, in case there are any issues that need to be worked out
first.

Fixes SR-12834.