ModuleLoading: account for standard library in load path
The previous change did fix the handling for non-standard library cases,
but the standard library path still required the old layout.  This
unifies the behaviour across the standard library and any other imported
libraries.  With this patch, it is almost possible to enable the new
layout in the build tree for Windows/Linux.