[NFC] Abort when PrintAsObjC hits a bad type
Currently, when DeclAndTypePrinter tries to print a type it doesn’t support, it fails very quietly by printing the Swift type in a comment. This almost certainly creates invalid Objective-C syntax, but that won’t be discovered until the header is imported, which is a Bad Thing.

This commit causes it to instead dump the offending type and abort with instructions to file a bug report.