Ban classes conforming to _ObjectiveCBridgeable
This commit adds a special rule to the declaration checker which makes it reject attempts to conform classes to _ObjectiveCBridgeable. It doesn’t make much sense to bridge a class to a different class when it comes into Swift, and PrintAsObjC has never handled these conformances correctly.

Banning is an aggressive solution, but we don’t promise source compatibility with underscored declarations, so it’s a matter of how badly this breaks things in practice. If this causes too much pain, we can instead add appropriate test coverage and fix PrintAsObjC.

Fixes rdar://62890964.