[builtins] Fix a broken build by removing a ScalableVecArgument case.
The following upstream commit removed the ScalableVecArgument:
https://reviews.llvm.org/D80107

I've remove that case from IntrinsicTypeDecoder::decodeImmediate()

The aforementioned patch also describes vector widths in terms of
ElementCounts, and not unsigned ints.  I've updated an access to the
vector width to make use of the ElementCount representation via access
of the Min member.