test: add a ODR violation check for the static standard library
The static version of the standard library was leaking symbols in the
`llvm::` namespace which would result in ODR violations were the
artifact linking against `LLVMSupport` (via another dependency).  In
particular, `llvm::SmallVector` and `llvm::StringSwitch` symbols were
being leaked.  This adds a test case specifically for the static variant
of the library.  The dynamic variant of the library is already tested in
a separate test.