[cmake] Move ar linking optimization rule into UnixCompileRules.cmake and only run it on Unix-like platforms.
This doesn't make sense on Windows and from what compnerd has said it doesnt
even appear in the ninja rules file. This also moves this rule out of the main
flow of the top level CMakeLists.txt file so I can reuse it when compiling the
stdlib using a sub-cmake invocation.