Add a Test Demonstrating Superfluous Cascading Across Module Boundaries
The idea is this: If we can have private intra-module dependencies, why
not also have private inter-module dependencies. Currently, this is
impossible because all edges from the CompilerInstance-level
DependencyTracker are unconditionally registered as cascading. In
addition, module loads are charged as depenendents to all files in
a given batch, which means the exact set of external dependencies is
some (non-deterministic) superset of the external dependencies of each
files in the batch.

This test case demonstrates one outcome of this approach: we recompile
*way* too much when a dependent module changes.